refactor: extract app-layout logic into reusable mixin
